refactor: Update DosenPembimbingLapanganController to use DataTable for JSON response
This commit modifies the DosenPembimbingLapanganController to use the DataTable helper class for generating the JSON response in the json() method. It introduces a search functionality that filters the data based on the provided search value. It also adds ordering capability based on the selected column and direction. These changes improve the functionality and performance of the JSON response in the DosenPembimbingLapanganController.