refactor: Update sidebar.blade.php to improve menu highlighting and add new menu items
This commit updates the sidebar.blade.php file to improve the highlighting of the active menu item. It replaces the hard-coded 'active current-page' class with a dynamic class based on the current route. Additionally, it adds new menu items for 'Mitra (Staff)', 'Dosen Pembimbing (Staff)', and 'Lowongan (Mitra)'. These changes enhance the clarity and functionality of the sidebar menu.