refactor: Add Peserta(Staff) menu item to sidebar
This commit adds a new menu item for Peserta(Staff) to the sidebar in the sidebar.blade.php file. The menu item is linked to the peserta.index route and displays an icon and text. This change enhances the navigation experience for staff members and improves the usability of the application.