refactor: Update migrations and controllers to add nullable fields
This commit updates the migrations and controllers to add nullable fields for certain columns. In the 'create_peserta_mbkm_table' migration, the 'jurusan', 'tanggal_lahir', 'jenis_kelamin', 'alamat', and 'telepon' columns are modified to be nullable. Similarly, in the 'create_dosen_pembimbing_lapangan_table' migration, the 'name', 'image', 'address', 'phone', 'email', and 'nip' columns are made nullable. This change allows for more flexibility in the data entry process by allowing these fields to be optional.