refactor: Add Dosen Pembimbing column to peserta list view
This commit adds a new column for "Dosen Pembimbing" to the peserta list view in the staff/registrasi-program/peserta/list.blade.php file. The column displays the name of the Dosen Pembimbing if the registration status is "placement" and the Dosen Pembimbing is assigned. Otherwise, it displays a dash ("-"). This change enhances the visibility of the Dosen Pembimbing information for each peserta in the list view.