refactor: Update MitraProfileController to store user_id when creating a new MitraProfile
This commit updates the MitraProfileController's store method to store the user_id when creating a new MitraProfile. The user_id is retrieved from the request and used to associate the MitraProfile with the corresponding user. This change ensures that the user is properly linked to the created MitraProfile and allows for better management of user-MitraProfile relationships.