refactor: Update AktivitasMbkmController.php to improve laporan harian creation process
The code changes in AktivitasMbkmController.php modify the process of creating a laporan harian. The changes include updating the 'mitra_id' field to be fetched from the 'registrationPlacement' relationship of the user's peserta, and removing unnecessary code related to the 'aktivitas' variable. This refactor improves the efficiency and accuracy of the laporan harian creation process.