refactor: Update BatchMbkmController to use validated data in store, update, and destroy methods
This refactor modifies the BatchMbkmController.php file to use the validated data from the request in the store, update, and destroy methods. Instead of using $request->all(), the code now uses $validatedData, which is obtained by validating the request data. This change improves the security and reliability of the data being stored, updated, and deleted in the BatchMbkm model.