refactor: Add About MBKM and Batch MBKM links to sidebar
This commit adds the "About MBKM" and "Batch MBKM" links to the sidebar in the sidebar.blade.php file. These links are added as list items with corresponding route links. This change enhances the navigation experience for users by providing easy access to the About MBKM and Batch MBKM pages.