refactor: Update BatchMbkm create and edit views
Update the create and edit views for BatchMbkm to include an "Active" checkbox field. This allows users to mark a batch as active or inactive. The "Active" field is added to the form and the value is stored in the database. The checkbox is disabled in the edit view to prevent modification of the active status. This change improves the user interface and provides better control over batch activation.