Refactor AktivitasMbkmController to optimize eager loading and fix peserta_id bug
The recent code changes in AktivitasMbkmController involve refactoring the index method to optimize eager loading of relationships and fix a bug related to the peserta_id parameter. The relationships 'registrationPlacement.lowongan.mitra' and 'registrationPlacement.dospem' are now eagerly loaded to avoid the N+1 problem. Additionally, the bug causing incorrect retrieval of peserta_id has been fixed. These changes improve the performance and reliability of the code.