feat: Improve questionnaire form UI and functionality
Refactor the `create.blade.php` file in the `applications/mbkm/questionnaire` directory to improve the UI and functionality of the questionnaire form. The changes include adding Bootstrap classes to enhance the styling and layout of the form elements. Additionally, the code is refactored to use the `form-control` and `form-select` classes for input fields and select dropdowns, respectively. These improvements enhance the user experience and make the form more visually appealing.

Co-authored-by: RahmatRafiq