Refactor QuestionnaireController.php to add showQuestionnaire method
This commit refactors the QuestionnaireController.php file to add a new method called showQuestionnaire. This method is responsible for displaying the questionnaire page and handling the logic to retrieve the list of participants who have filled out the questionnaire. The code that was commented out in the previous version of the file has been removed, and the showQuestionnaire method now retrieves the list of participants and returns the appropriate view. This change improves the functionality of the questionnaire feature and enhances the user experience.

Co-authored-by: RahmatRafiq