feat: Add Bossa font to the project
The Bossa font has been added to the project. This font is used for the 'sans' and 'bossa' font families in the Tailwind CSS configuration. It enhances the visual appeal of the website and provides a unique typography style. This change improves the overall design and aesthetics of the project.

Co-authored-by: RahmatRafiq